Refactor trace store test module
